codelab_docs/scripts/requirements.txt
google-genai>=0.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
